WIP Spark tf-idf

+++ src/main/java/com/bdea/grp2/lambda/model/Df.java
@@ -19,5 +19,5 @@ import javax.persistence.Table;
 public class Df {
     @Id
     private String term;
-    private long df;
+    private int df;
 }
